Global Hook Fixed, more project fixes
Set window hook changed in .net 4

Found out the original author was registering in the gac, I'm going to
remove that for now, if I need it later, I'll do it, until then I'm
happy leaving it out.

Fixed up the EntryPointDocumentView xaml, couldn't save, or a plethora
of other small issues.